Bug fix: Changed from static to dynamic thread struct allocation
The array thread structures that were being allocated statically (in main())
consumed a lot of space on the stack, which prevented the later
functions such as pcm_param() to be called. This was the case
with large PCM_ROWS_PER_BANK value, which consumed a lot of space
for the 'rows' array inside the thread structure.

The solution was to allocate memory for it dynamically.